bt-unit13

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,11 +19,11 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->tinyInteger('status')->default(1);
+            $table->integer('status')->default(1);
             $table->longText('avatar')->nullable();
             $table->char('phone')->nullable();
             $table->string('address')->nullable();
-            $table->tinyInteger('role')->default(0);
+            $table->integer('role')->default(0);
             $table->rememberToken();
             $table->timestamps();
         });
